fix(web): aggregate locale readiness in SQL and fix client bundle
Replace the 500-row locale readiness sample with a Postgres jsonb_each_text
aggregation so all provider files contribute to dashboard metrics.

Split types into a client-safe module so the dashboard section no longer
imports the database layer and breaks the Next.js build.
